refactor: cleanup client and server
